pyproject.toml
[project]
name = "weather-sales-predictor"
version = "0.1.0"
description = "Hackathon MVP for predicting retail sales from weather features."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "databricks-sdk>=0.67.0",
  "joblib>=1.4.2",
  "numpy>=2.2.5",
  "pandas>=2.2.3",
  "plotly>=6.0.1",
  "scikit-learn>=1.6.1",
  "streamlit>=1.44.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
